eden: add `eden du` command for inspecting disk usage
Summary:
Understanding disk usage and cleaning it up is a relatively frequently
asked question, and users are often shocked that "eden is using so much space"
when the reality is that they're just seeing the aggregate from a `du` command
that includes their build artifacts from their bind mounts.

This diff adds `eden du` as a friendly utility that helps to break down the
disk usage into meaningful categories and gives some hints on how to reclaim
space.

Reviewed By: chadaustin

Differential Revision: D17252322

fbshipit-source-id: 7348692c95cd319ca315c5be466b397b4969971d