mononoke: lfs_server: Remove bail_http macros
Summary:
Previously, the LFS server's internal functions returned HandlerResponses
which included HTTP status codes. The bail_http macros were used to implement this,
but made it difficult pass state to them for logging, as well as not allowing the use
of the ? operator, which is a nice benefit of async / await syntax.

Refactor these functions to that they don't return such Gotham specific types
and remove the relevant http macros.

Reviewed By: krallin

Differential Revision: D17286238

fbshipit-source-id: 05ff791d4761b0f742d22a2966d5ecc5968728ba