hgbuild: use sandcastle DISK_TEMP rather than scratch on sandcastle
Summary:
This is a short term workaround for an issue affecting sandcastle
where the scratch dir is being created as root and subsequently blocking
the job from creating new scratch dirs inside it.

We'll fix this up in some sandcastle specific config in a follow up diff,
but this should unblock the hgbuild for now.

Reviewed By: quark-zju

Differential Revision: D17296990

fbshipit-source-id: 823a86312a6684b385395bd6427c9b8ce2639d5e