rust/configerator: Require fbinit proof in ConfigeratorAPI
Summary:
Since ConfigeratorAPI::new requires initFacebook to have been called, this diff gives it a `Facebook` parameter to require that callers prove they have called #[fbinit::main] as described in D17245802.

This diff turned out to be not too invasive in my opinion. It turns out there are usually pretty few layers between main and ConfigeratorAPI::new. Threading along a `Facebook` object beats the experience of debugging an error that looks like this: P109696408 when forgetting to use fbinit.

Reviewed By: Imxset21

Differential Revision: D17277841

fbshipit-source-id: 2fe3096ebcac58bb123149906e7e5d9d9e2da685