revisionstore: rename IterableStore to ToKeys
Summary:
It's hard (impossible?) to write the iter method for a Store that uses interior
mutability with RefCell and/or Mutex. Due to how MutableDeltaStore and
MutableHistoryStore are designed, it's impossible to use a MutableDeltaStore in
a UnionStore while still being able to add to it due to the requirement of the
add method to be passed in a `&mut self`. To change this, the method needs to
be changed to accept a `&self`, which thus requires the use of a Mutex/RefCell.

Since the IterableStore is never used to get a subset of the keys, let's change
and rename it to always return a Vec of all the keys.

Reviewed By: quark-zju

Differential Revision: D17278150

fbshipit-source-id: 64bad54e69dd89a91acad0d0877d888685858994