mononoke: prefetch commits in backfill derive data
Summary:
Fetching commits takes a long time. I've added a subcommand that saves them to
a file, and "backfill" command can later read it. Note that another option
would be to add indices to the db to make commit fetching faster, however this diff
was simpler to do.

I also fixed a panic which was because of division by zero.

Reviewed By: aslpavel

Differential Revision: D17341640

fbshipit-source-id: b0335ebf8799cd48884c19fa8a0ee8023eb751af