manifest: add initial bfs diff implementation
Summary: This diff provides an implementation of the diff operation for trees which processes directories in BFS order (i.e., layer by layer). This allows the iterator to perform a bulk prefetch of the changed nodes in each layer at the start of each layer of the traversal. This should hopefully provide a more efficient fetch pattern than the existing implementation, which requires a full prefetch of both trees upfront for reasonable performance.

Reviewed By: xavierd

Differential Revision: D17276971

fbshipit-source-id: 284f1d458f43cb76befe27e85f53a641f29d7550