mononoke: use spawn_future in fastlog generation
Summary:
Generating a lot of fastlog batches for a single commit turned out to be a cpu
intensive operation. To make sure more cpu threads are used let's use
spawn_future() function that puts each future in a separate Tokio task, which
in turn let it be scheduled on different cpu.

There are some concerns that it might cause problems when we'll derive data in
production i.e. the whole host will be unavailable for a few seconds because it
uses all cpus. My hope is that it will affect only very few hosts because of
memcache leases that we use in derive data implementation.

Reviewed By: krallin

Differential Revision: D17364581

fbshipit-source-id: 0281de9c42f93d793f6caccbca7e8809056782ec