remotefilelog: add the indexedlog to remotefilelog.shareddatastores
Summary:
Ondemand has been complaining that their hgcache was growing very quickly, and
looking at logs and scuba data, this appears to be caused by refetching
everything from the network and writing it to the indexedlogdatastore.

Looking at the prefetch code, I realized that the indexedlogdatastore wasn't
present in the repo.fileslogs.shareddatastore list, and therefore may be
omitted from the stores in which we look for data.

Reviewed By: kulshrax

Differential Revision: D17374761

fbshipit-source-id: 9e9f279d4c8154e2491312a57d24bc5fd7da79fc