dag: add a flag storing whether a segment has roots
Summary:
This would allow us to implement the fast path in `children`.

This is a breaking change in file format. But we haven't been really using the
segmented dag format in production yet. So it's fine to not have a migration
path.

This potentailly makes the future format migration easier since there are 7
bits left to use.

Reviewed By: markbt

Differential Revision: D16976382

fbshipit-source-id: 0110c03d4ff2ce0a882b3fb8ce8ee12d2087f41f