mononoke: Update multiplexedblob logging to include key and fix error logging
Summary:
Change the multiplexed blobstore logging to log the blobstore key.

Also, fix the get() method to log errors to scuba (which it wasn't before!).

Further, refactor the code to remove env parsing for tupperware information and also switch
to using ScubaSampleBuilder. This means that the common server information can
be stored once and then cloned each time we log, rather than iterating through
a vec of server information each time we want to log.

Additionally, using ScubaSampleBuilder means that we don't need to pass in an Option<ScubaClient>,
cleaning up the code a little bit.

Reviewed By: StanislavGlebik

Differential Revision: D17368779

fbshipit-source-id: 0896962cdbd37912fc6f23a5e541e10cea90fa0e