mkscratch: add $REPO_OWNER_HOME
Summary:
$HOME and $USER can vary depending on whether sudo is used or
not.  In the mixed owner situations that tend to crop up at FB we
generally want to consider the user and home dir of the owner of the
repo rather than whatever the current user is.

This diff adds a `REPO_OWNER_HOME` replacement token that resolves
the owner of the repo and then looks up their home directory.

Reviewed By: simpkins

Differential Revision: D17335435

fbshipit-source-id: bf81565abdf66084b93c3411ae6b58af0c4a6143