mkscratch: if run as root, chown files/dirs to the repo owner user
Summary:
If we are run as root then set the owner of any files or
directories that we create to the owner of the repo.

This helps to avoid some classes of problems if there are mixed
invocations with root and repo owner.

Note that I picked a string to represent the user largely because
I can see us wanting to specify the user in the config file
using a string at some point in the future.

It is also slightly more portable (even though it isn't native)
than doing something funky with uid/gid on posix and something
involving security descriptors or DACLs on windows.

Reviewed By: chadaustin

Differential Revision: D17335433

fbshipit-source-id: 33c1d1cdbe4d07ce4d4f2581206f1308e3b29290