revisionstore: add Mutable{HistoryPackStore,DataPackStore}
Summary:
Similarly to the IndexedLog based store, let's also enable writing to the
packstore. One big benefit of doing that is that it will greatly reduce the
need for rescanning the filesystem when new packfiles are created.

The read-only version are for now kept as they are being exposed to Python,
long term, the read-only will be made private and the Mutable version renamed.

Reviewed By: kulshrax

Differential Revision: D17357937

fbshipit-source-id: 0f440d5668ca8b412300ff3519a233d1449a464d