getdeps: restructure bison+flex dependencies
Summary:
We've been squeaking by with assuming that flex is installed already
on posix systems, but that isn't the case on the github actions default
configuration.

Adjust the bison recipe: on windows it deploys both flex and bison.  We use the
same source for both flex and bison but install flex to a separate install
prefix to make it easier to consume the flex dependency distinct from the bison
dependency.

The latest flex release segfaults during compilation on linux unless we
force -DGNU_SOURCE, so the manifest does that on linux.

Reviewed By: simpkins

Differential Revision: D17385051

fbshipit-source-id: 9f31b07849af9de50099d1b20bedba517bbbdf2f