Sync commits between two repos
Summary: To make merges into our final megarepo world Just Work, we want to sustain an interim period where the legacy repo is still present and up to date. Provide a way to do this by having Mononoke internally remap and pushrebase commits so that we can have two repos kept in sync.

Reviewed By: ikostia, StanislavGlebik

Differential Revision: D16918280

fbshipit-source-id: c7a8d1bb1b1972cb5609490b8fd5b759cb5f9cdd