mononoke/multiplexblob: don't wait for SQL Queue if all writes have succeeded
Summary:
This updates multiplexblob to avoid waiting for the SQL queue if writes have
succeeded in all the blobstores we are writing to.

It's worth noting that this might mean adding a new blobstore will require a
little more effort, since we no longer guarantee that the blobstore queue will
contain all writes that we made. That said, we weren't really in a position to
100% rely on this anyway (writes can in fact succeed even if the SQL queue
doesn't get written to), so in that sense it doesn't make a huge difference.

Reviewed By: farnz

Differential Revision: D17421208

fbshipit-source-id: 0f2ecbf22ba51531a5917baf912183d5309e1b63