mononoke: lfs_server: Store a scuba logger in State
Summary:
Store a wrapped instance of ScubaSampleBuilder in gotham's State, so that it can be
accessed from outside of the middleware. Further, refactor the Scuba middleware
to use a separate function and replace the and_then() combinator with then() so that all internal
errors are handled by the scuba logger.

Additionally, use the scuba logger to log the number of objects in a batch as well
as the size of uploads.

Reviewed By: krallin

Differential Revision: D17425371

fbshipit-source-id: bb86995c2e561062c1b1951fcea98fa25300103c