eden: du: ignore FileNotFoundError when stat'ing ignored files
Summary:
It's possible that status includes ignored files from
the overlay that are no longer accessible because a bind mount
is overlaid on top of the overlay!

Ideally we'd account for the space occupied by those files,
but for now we'd rather not error out when attempting to stat
them, so silently ignore that class of error.

Reviewed By: chadaustin

Differential Revision: D17415965

fbshipit-source-id: 96d467ee4e85ad5bdc730e4cfffe4c15a83a12ed