fix a crash in GlobNode::evaluate()
Summary:
Fix `GlobNode::evaluate()` so that it only generates a result once all
processing is complete.  The glob evaluation walks down the inode tree in
parallel.  Previously if an error occurred while processing one TreeInode it
would return that exception to the caller immediately, even if work was still
being done to evaluate the glob on other parts of the inode tree.  This
behavior is problematic, since the caller may destroy the GlobNode immediately
once a result is generated, and the caller has no way to know that other
children GlobNodes may still be using some of the data.

This fixes the code to wait until all evaluation sub-futures have completed
processing even on error, so that we know it is safe to destroy the GlobNode
before we generate a result.

Reviewed By: wez

Differential Revision: D17434327

fbshipit-source-id: ef70dc82281eeb3540840202ec5e704afe2d29eb