edenapi: remove dependency on revisionstore
Summary:
The revisionstore crate will soon need to build a `dyn EdenApi` object to fetch
data out of the network. Since the edenapi crate depends on revisionstore, this
would create a recursive dependency between these 2. Several approach were
thought of, including moving either the EdenApi, or the MutableDeltaStore trait
outside of their respective crates. In the end, I decided to remove the
dependency altogether and let the caller decide what to do about the data.

Reviewed By: quark-zju

Differential Revision: D17437895

fbshipit-source-id: cc3ec830562c0616d40d7d5d36f69674934d87b9