revisionstore: add a ContentStore struct
Summary:
The ContentStore struct is intended to be the main entry point for data and
cache management for both Mercurial and Eden. It takes care of finding the
proper underlying stores and aggregate them into one, this is roughly
equivalent to the repo.datastore object in Python, the main different being
that the Rust one isn't yet fetching any remote data. This will be added in a
future diff.

Reviewed By: quark-zju

Differential Revision: D17357935

fbshipit-source-id: a47575177cbb7315ea22e3da9f4043845698c61f