manifest: add sort order edge case test for diff
Summary:
The diff algorithm assumes that the cursor iterates over the entries of the
manifest in order. The problem is that a tree manifest with sorted names
for the entries of the directories does not yield the same order as the
string sorted paths. The path separator: '/' is the problem. It is greater
than certain charater, like '-'. This test validates that the diff algorithm
works for manifests that would not yield string sorted order paths on DFS
iteration.

Reviewed By: quark-zju

Differential Revision: D17491659

fbshipit-source-id: 0754a19676ddd6500ea11b07446936d922b99652