eden: fixup mononoke fetching when using the VIP
Summary:
I was experimenting with using direct mononoke fetches from my mac but
was hitting 400 errors.  It turned out that we were resolving the VIP name and
using the resolved host in the generated URL.  The VIP configuration doesn't
know how to route such a request and yields a 400 response.

The SocketAddressWithHostname type has enough information to allow for handling
this correctly, but we were discarding the hostname component!

This diff captures the entire SocketAddressWithHostname type in the `address_`
member and uses the hostname portion when generating the URL.

Reviewed By: fanzeyi

Differential Revision: D17470256

fbshipit-source-id: 14ebd517eac12e07248a64fb37c0623ecbe37e89