lfs: add test for lfs-downgrade-after-rename issue
Summary:
Normally, when you push an lfs blob the server doesn't need to ever
download the lfs blob content because it can just compare the hashes. There's a
special case though, when a commit makes a blob go from lfs to non-lfs and the
previous version of the blob was a rename, then Mercurial must actually read the
file contents and compare it to know if it changed.

This adds a test to cover the case.

Reviewed By: quark-zju

Differential Revision: D17488079

fbshipit-source-id: 4313ee01e103d3ad13463e4774096d866d420b2e