indexedlog: avoid mem::swap which breaks Log internal state
Summary:
There are several places in Log that use `mem::swap` to swap out existing
Indexes, with the expectation that self.indexes will be replaced with updated
indexes soon. But that can fail and Log will be in a bad state.

Fix that by replacing `mem::swap` with clones so it is easy to verify that
the internal state of `Log` wouldn't be broken.

The extra `clone` makes certain `sync` operations a bit slower:

Before:

  log sync (write N)                                 31.155 ms      20 B /  9.259 MB
  log sync (read, new log 1, index lag N)             0.033 ms      42 B /      0 B

After:

  log sync (write N)                                 39.841 ms      20 B /  9.259 MB
  log sync (read, new log 1, index lag N)             3.343 ms      42 B /      0 B

This is caught by an assertion error at log.rs:1095, reported by fanzeyi.

Reviewed By: xavierd

Differential Revision: D17493796

fbshipit-source-id: 7e8b4329258ca4d24be74aa942ffa809397b2eaf