sampling: add a debug option
Summary:
This config prints out data written by the sampling extension.

The motivation behind this is, we want the Rust code to also be able to
generate data that can be sent to our remote logging infra (scuba).
Right now, the only way to do that is to go through sampling.py, and
use the `ui.log` API. I'm looking to change it to use Rust code and
blackbox data instead. To get some confidence about the upcoming Rust
code, I need a way to easily compare data using production setup.

Reviewed By: singhsrb

Differential Revision: D17494927

fbshipit-source-id: 0b10562925ec82dbd67f18996d7bfeab8cfd81ee