hgcommands: log process start event
Summary:
This will provide the "fullargs" information needed by telemetry.

Log in "hgcommands", instead of "hgmain", or "clidispatch", because:
- "clidispatch" does not have  argv[0], and might be called multiple times
  (from Rust, then from the fallback Python -> bindings)
- Logs in "hgmain" won't be written to disk if chg is used.

I removed "env", since "hgcommands" does not have access to hgrc to decide what
env to log. It seems desirable if the "env" logging is move to a fb-only
backgroud logging binary.

Reviewed By: singhsrb

Differential Revision: D17423796

fbshipit-source-id: da3c54e916203f8368f8389fc7eed91e5299a75d