mononoke: move find_new_unodes to ManifestOps from fastlog
Summary:
This function finds entries that were "introduced" by this manifest i.e.
entries that are present in this manifest but not present in any of
`diff_against` manifests.

This function was used in fastlog, and I'm planning to use it in the new sync
job, so move it to ManifestOps crate (and rename along the way).

One small change in behaviour is that we removed a single traced...() call. I
talked with aslpavel who introduced it and he's fine with removing it -
unfortunately our traces are not in a great shape yet, so this trace call is
not used at all.

Reviewed By: aslpavel

Differential Revision: D17477055

fbshipit-source-id: cba1a55a64299857efe4f4be7a67b6faf31e7019