mononoke/lfs_server: don't wait for upstream if we can respond locally
Summary:
If we have all the objects a client is asking for, we might as well answer
immediately with them, rather than wait for upstream.

This effectively means that this LFS server can be faster than its upstream on
batch downloads (for uploads, we still have to check in with upstream), as it's
no longer limited by upstream's performance (or errors!) when responding to
batch requests.

Reviewed By: HarveyHunt

Differential Revision: D17526491

fbshipit-source-id: 5d008011fd5ac444c84569f8235ecc9bde6cef9c