copytrace: chose between gdbm and anydb for the amendcopytrace db
Summary:
The fbcode runtime Python unfortunately isn't compiled with the _bsddb module
in, and thus when using the packman built Mercurial binary, renaming a file and
performing an amend will cause it to crash. Unfortunately, we can't simply use
gdbm everywhere as Windows doesn't support it, so let's hack around this by
trying both anydbm and gdbm.

Reviewed By: quark-zju

Differential Revision: D17512253

fbshipit-source-id: 3717b691e715084b7c3488425ae20dcf7e9ad2bd