mononoke: first diff for the sync job that generates bundles (only single commit)
Summary:
This is the very first and very limited implementation of a new mode of the
sync job. This is the mode where bundles are generated by the sync job instead
of reusing the bundles pushed by the client.

At the moment it syncs only a single commit, new functionality has to be added
later.

A few notes:
1) We need to manually generate replycaps part . Mononoke essentially ignores it
(with a few small exceptions), however Mercurial actually pays attention to it, and
generates different response.
At the moment reply capabilities are hard-coded in the sync job, and that might be
problematic if replycaps change by Mercurial client but sync job hasn't been
updated.
There are a few protections from it - we can make replycaps dynamic (for
example, specify it in the tw spec) and we can also rely on integration tests
to catch regressions.

2) create_filenode_entries_stream has a complicated logic for renames - please
have a look at it to double check my understanding!

Reviewed By: krallin

Differential Revision: D17477983

fbshipit-source-id: fdb4584e768bdc5637868468a035c81f1584f7fe